Refactor #68 - move `ScheduleUiStates` package util to top
