+++ app/src/main/java/com/nlab/reminder/domain/feature/schedule/all/AllScheduleStateMachine.kt
@@ -24,7 +24,7 @@ import com.nlab.reminder.core.kotlin.util.onFailure
 import com.nlab.reminder.core.state.StateMachine
 import com.nlab.reminder.domain.common.schedule.*
 import com.nlab.reminder.domain.common.schedule.SelectionModeRepository
-import com.nlab.reminder.domain.common.schedule.util.asSelectedSchedules
+import com.nlab.reminder.domain.common.schedule.asSelectedSchedules
 import com.nlab.reminder.domain.common.schedule.visibleconfig.*
 
 /**
